requirements.txt
Flask==3.0.0
pycryptodome==3.19.0
qrcode==7.4.2
Pillow==10.1.0
